Extract display type to separate class
