+++ src/main/java/com/cubefury/vendingmachine/blocks/gui/TradeItemDisplayWidget.java
@@ -1,8 +1,5 @@
 package com.cubefury.vendingmachine.blocks.gui;
 
-import static com.cubefury.vendingmachine.gui.GuiTextures.MODE_LIST;
-import static com.cubefury.vendingmachine.gui.GuiTextures.MODE_TILE;
-
 import net.minecraft.item.ItemStack;
 
 import org.jetbrains.annotations.NotNull;
@@ -12,8 +9,6 @@ import com.cleanroommc.modularui.api.value.IValue;
 import com.cleanroommc.modularui.api.widget.Interactable;
 import com.cleanroommc.modularui.drawable.DynamicDrawable;
 import com.cleanroommc.modularui.drawable.GuiDraw;
-import com.cleanroommc.modularui.drawable.Icon;
-import com.cleanroommc.modularui.drawable.UITexture;
 import com.cleanroommc.modularui.screen.viewport.ModularGuiContext;
 import com.cleanroommc.modularui.theme.WidgetThemeEntry;
 import com.cleanroommc.modularui.utils.Platform;
@@ -26,29 +21,6 @@ import com.cubefury.vendingmachine.gui.WidgetThemes;
 import com.cubefury.vendingmachine.util.Translator;
 
 public class TradeItemDisplayWidget extends ItemDisplayWidget implements Interactable {
-
-    public enum DisplayType {
-
-        TILE("tile", MODE_TILE),
-        LIST("list", MODE_LIST);
-
-        private final String type;
-        private final Icon texture;
-
-        DisplayType(String type, UITexture texture) {
-            this.type = type;
-            this.texture = texture.asIcon();
-        }
-
-        public String getLocalizedName() {
-            return IKey.lang("vendingmachine.gui.display_mode_" + this.type)
-                .toString();
-        }
-
-        public Icon getTexture() {
-            return this.texture;
-        }
-    }
 
     private MTEVendingMachine vm;
     private TradeMainPanel rootPanel;
